pyproject.toml
[tool.poetry]
name = "Retrieval_docs_agent"
version = "0.1.0"
description = ""
authors = ["ehdtjr <dskang207@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohappyeyeballs = "2.5.0"
aiohttp = "3.11.13"
aiosignal = "1.3.2"
annotated-types = "0.7.0"
anyio = "4.8.0"
appnope = "0.1.4"
asttokens = "3.0.0"
attrs = "25.1.0"
beautifulsoup4 = "4.13.3"
certifi = "2025.1.31"
charset-normalizer = "3.4.1"
click = "8.1.8"
comm = "0.2.2"
cssselect = "1.3.0"
dataclasses-json = "0.6.7"
debugpy = "1.8.13"
decorator = "5.2.1"
dependency-injector = "4.46.0"
distro = "1.9.0"
et_xmlfile = "2.0.0"
executing = "2.2.0"
fastapi = "0.115.11"
feedparser = "6.0.11"
frozenlist = "1.5.0"
goose3 = "3.1.19"
h11 = "0.14.0"
httpcore = "1.0.7"
httpx = "0.28.1"
httpx-sse = "0.4.0"
idna = "3.10"
ipykernel = "6.29.5"
ipython = "9.0.2"
ipython_pygments_lexers = "1.1.1"
jedi = "0.19.2"
jiter = "0.8.2"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
jupyter_client = "8.6.3"
jupyter_core = "5.7.2"
langchain = "0.3.20"
langchain-community = "0.3.19"
langchain-core = "0.3.41"
langchain-openai = "0.3.7"
langchain-text-splitters = "0.3.6"
langdetect = "1.0.9"
langgraph = "0.3.5"
langgraph-checkpoint = "2.0.16"
langgraph-prebuilt = "0.1.1"
langgraph-sdk = "0.1.53"
langsmith = "0.3.11"
lxml = "5.3.1"
markdown-it-py = "3.0.0"
marshmallow = "3.26.1"
matplotlib-inline = "0.1.7"
mdurl = "0.1.2"
msgpack = "1.1.0"
multidict = "6.1.0"
mypy-extensions = "1.0.0"
nest-asyncio = "1.6.0"
numpy = "2.2.3"
openai = "1.65.3"
openpyxl = "3.1.5"
orjson = "3.10.15"
outcome = "1.3.0.post0"
packaging = "24.2"
pandas = "2.2.3"
parso = "0.8.4"
pexpect = "4.9.0"
pillow = "11.1.0"
pip = "25.0.1"
platformdirs = "4.3.7"
prompt_toolkit = "3.0.50"
propcache = "0.3.0"
psutil = "7.0.0"
psycopg2-binary = "2.9.10"
ptyprocess = "0.7.0"
pure_eval = "0.2.3"
pyahocorasick = "2.1.0"
pydantic = "2.10.6"
pydantic_core = "2.27.2"
pydantic-settings = "2.8.1"
Pygments = "2.19.1"
PySocks = "1.7.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2025.1"
PyYAML = "6.0.2"
pyzmq = "26.3.0"
regex = "2024.11.6"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rich = "13.9.4"
ruff = "0.9.9"
selenium = "4.30.0"
setuptools = "65.5.0"
sgmllib3k = "1.0.0"
six = "1.17.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
soupsieve = "2.6"
SQLAlchemy = "2.0.38"
stack-data = "0.6.3"
starlette = "0.46.1"
tavily-python = "0.5.1"
tenacity = "9.0.0"
tiktoken = "0.9.0"
tornado = "6.4.2"
tqdm = "4.67.1"
traitlets = "5.14.3"
trio = "0.29.0"
trio-websocket = "0.12.2"
typing_extensions = "4.12.2"
typing-inspect = "0.9.0"
tzdata = "2025.2"
urllib3 = "2.3.0"
uvicorn = "0.34.0"
wcwidth = "0.2.13"
webdriver-manager = "4.0.2"
websocket-client = "1.8.0"
wsproto = "1.2.0"
yarl = "1.18.3"
zstandard = "0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
